refactoring events models and tests file structure
