+++ hknweb/events/models/__init__.py
@@ -1,0 +1,3 @@
+from hknweb.events.models.event_type import EventType
+from hknweb.events.models.event import Event
+from hknweb.events.models.rsvp import Rsvp
