+++ hknweb/events/models/event.py
@@ -5,17 +5,7 @@ from markdownx.models import MarkdownxField
 
 from hknweb.utils import get_semester
 
-
-class EventType(models.Model):
-    type = models.CharField(max_length=255)
-    # Default color: CS61A blue
-    color = models.CharField(max_length=7, default="#0072c1")
-
-    def __repr__(self):
-        return "EventType(type={})".format(self.type)
-
-    def __str__(self):
-        return str(self.type)
+from hknweb.events.models.event_type import EventType
 
 
 class Event(models.Model):
@@ -35,15 +25,8 @@ class Event(models.Model):
         ],
         default=0,
     )
-    # need_transportation = models.BooleanField(default=False)
-    # view_permission_group_id = models.IntegerField(null=True)
-    # rsvp_permission_group_id = models.IntegerField(null=True)
-    # view_permission_group_id = models.ForeignKey('people.Group', on_delete=models.SET_NULL, null=True)
-    # rsvp_permission_group_id = models.ForeignKey('people.Group', on_delete=models.SET_NULL, null=True)
-    # markdown    = models.BooleanField(default=False)
     created_by = models.ForeignKey(User, on_delete=models.CASCADE, default=None)
     created_at = models.DateTimeField(auto_now_add=True)
-    # updated_at  = models.DateTimeField(auto_now=True)
 
     @property
     def semester(self):
@@ -83,24 +66,3 @@ class Event(models.Model):
         """old_admitted must be a set, not a QuerySet. QuerySets are mutable views into the database."""
         new_admitted = set(self.admitted_set())
         return new_admitted - old_admitted
-
-
-class Rsvp(models.Model):
-    user = models.ForeignKey(User, models.CASCADE, verbose_name="rsvp'd by")
-    event = models.ForeignKey(Event, models.CASCADE)
-    confirmed = models.BooleanField(default=False)
-    comment = models.TextField(blank=True, default="")
-    # transportation  = models.IntegerField(choices=TRANSPORT_ENUM,
-    #                                       default=HAVE_RIDE)
-    created_at = models.DateTimeField(auto_now_add=True, verbose_name="rsvp time")
-    # updated_at      = models.DateTimeField(auto_now=True)
-
-    def __repr__(self):
-        return "Rsvp(event={}, user={})".format(self.event, self.user.username)
-
-    def __str__(self):
-        return self.event.name
-
-    @classmethod
-    def has_not_rsvpd(cls, user, event):
-        return cls.objects.filter(user=user, event=event).first() is None
